see details
change newindex to index, thanks to xice's work.
add the notice about the scan qrcode
add a recursive  func "scanQR"
