+++ client/pages/index/index.json
@@ -1,3 +1,1 @@
-{
-  "enablePullDownRefresh": true
-}
+{}
