+++ client/pages/index/index.wxml
@@ -1,16 +1,49 @@
-<!--index.wxml-->
 <template name="fileList">
-  <view class='list-item' bindtap='showDetails' data-type="{{type}}" id="{{id}}" data-id="{{id}}" data-name="{{fileName}}" data-time='{{uploadTime}}'>
-    <image src="https://asdf.zhr1999.club/resource/icons/{{type}}.svg" class='file-icon'></image>
-    <view class='text-box'>
-      <text class='file-name'> {{fileName}} </text>
-      <text class="file-time"> {{uploadTime}}</text>
+  <block>
+    <view class='list-item' bindtap='showDetails' data-type="{{type}}" id="{{id}}" data-id="{{id}}" data-name="{{fileName}}" data-time='{{uploadTime}}'>
+      <image src="https://asdf.zhr1999.club/resource/icons/{{type}}.svg" class='file-icon'></image>
+      <view class='text-box'>
+        <text class='file-name'> {{fileName}} </text>
+        <text class="file-time"> {{uploadTime}}</text>
+      </view>
+    </view>
+  </block>
+</template>
+<template name="group-list">
+  <block>
+    <view class='list-item' bindtap='showDetails' data-type="group" id="{{id}}" data-id="{{id}}">
+      <image src="https://asdf.zhr1999.club/resource/icons/group.svg" class='file-icon'></image>
+      <view class='text-box'>
+        <open-data type="groupName" open-gid="{{id}}" />
+      </view>
+    </view>
+  </block>
+</template>
+<view>
+  <scroll-view scroll-x="true" class="tab-h" scroll-left="{{scrollLeft}}" scroll-with-animation="{{true}}">
+    <view class="tab-item {{currentTab==0?'active':''}}" data-current="0" bindtap="swichNav">我的文件</view>
+    <view class="tab-item {{currentTab==1?'active':''}}" data-current="1" bindtap="swichNav">群组</view>
+  </scroll-view>
+  <view class="refresh-block" style="  margin-top: {{refreshHeight}}px;;" wx:if="{{refreshing}}">
+    <image class="loading-gif" src="https://asdf.zhr1999.club/resource/pngs/loading.gif"></image>
+  </view>
+  <swiper class="tab-content" current="{{currentTab}}" duration="300" bindchange="switchTab" style="height:{{winHeight}}px">
+    <swiper-item>
+      <scroll-view class="container" scroll-y="true" style="padding:0rpx;height:100%" wx:if="{{loginStatus}}" pper-threshold="0" bindscrolltoupper="toUpperLoad">
+        
+        <template is="fileList" wx:for="{{filelist.data}}" wx:key="id" wx:if="{{!filelist.empty}}" data="{{...item}}"></template>
+      </scroll-view>
+    </swiper-item>
+    <swiper-item>
+      <scroll-view class="container" scroll-y="true" style="padding:0rpx;height:100%" wx:if="{{loginStatus}}" pper-threshold="0" bindscrolltoupper="toUpperLoad">
+        <template is="group-list" wx:for="{{groupList.group}}" wx:key="id" wx:if="{{!groupList.empty}}" data="{{...item}}"></template>
+      </scroll-view>
+    </swiper-item>
+
+  </swiper>
+  <view class="com-widget-scanCode" bindtap="scanCode">
+    <view class="icon-scanCode">
+    <text class="fa fa-upload"></text>
     </view>
   </view>
-</template>
-<view class="container" style='padding:15rpx' wx:if="{{loginStatus}}">
-  <view class="userinfo" wx:if="{{!hasUserInfo && canIUse}}">
-    <button open-type="getUserInfo" bindgetuserinfo="getUserInfo"> 获取用户信息 </button>
-  </view>
-  <template is="fileList" wx:for="{{filelist.data}}" wx:key="id" wx:if="{{!filelist.empty}}" data="{{...item}}"></template>
 </view>
