--- client/pages/newIndex/newIndex.json
@@ -1,1 +1,0 @@
-{}
