--- client/pages/newIndex/newIndex.wxss
@@ -1,134 +1,0 @@
-.tab-content {
-  margin-top: 90rpx;
-}
-
-.tab-h {
-  height: 90rpx;
-  width: 100%;
-  box-sizing: border-box;
-  overflow: hidden;
-  line-height: 80rpx;
-  background: #51A8DD;
-  font-size: 16px;
-  white-space: nowrap;
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 99;
-}
-
-.tab-item {
-  width :50%;
-  text-align: center;
-  font-size: 105%;
-  color: #fff;
-  display: inline-block;
-}
-
-.tab-item.active {
-  position: relative;
-}
-
-.tab-item.active:after {
-  content: "";
-  display: block;
-  height: 8rpx;
-  width: 50%;
-  background: #fff;
-  position: absolute;
-  bottom: 0;
-  top: 85rpx; 
-  left: 25%;
-  right: 25%;
-  border-radius: 16rpx;
-}
-.userinfo {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.userinfo-avatar {
-  width: 128rpx;
-  height: 128rpx;
-  margin: 20rpx;
-  border-radius: 50%;
-}
-
-.userinfo-nickname {
-  color: #aaa;
-}
-
-.usermotto {
-  margin-top: 200px;
-}
-image.file-icon{
-  height: 140rpx;
-  width: 140rpx;
-  margin: 15rpx 20rpx 15rpx
-}
-.loading-gif{
-  height: 20rpx;
-  width: 100rpx;
-  margin-top: 90rpx;
-}
-.list-item{
-  border:1px solid #eeeeee;
-  width: 96%;
-  display: flex;
-  flex-direction: row;
-  padding: 1%;
-  margin-bottom: 15rpx;
-  margin-left: 1%;
-  margin-right: 2%;
-  box-shadow: 2px 2px 2px #eeeeee;
-}
-
-view.text-box{
-  display:flex;
-  flex-direction:column;
-  justify-content: space-around;
-  align-items: flex-start;
-}
-
-text.file-name{
-  font-size: 36rpx;
-  margin-top:-15rpx;
-
-}
-
-text.file-time{
-  font-size: 20rpx;
-  color: #bdbdbd;
-}
-.refresh-block{
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  height: 50px;
-}
-.com-widget-scanCode {
-        position: fixed;
-        bottom: 75px;
-        right: 20px;
-        background: #51A8DD;
-        border-radius: 50%;
-        overflow: hidden;
-        z-index: 500;
-        border:1px solid #eeeeee;
-        box-shadow: 2px 2px 2px #eeeeee;
-}
-.com-widget-scanCode .icon-scanCode{
-        background-color: #51A8DD;
-        display: inline-block;
-        width: 50px;
-        height: 50px;
-        line-height: 50px;
-        font-size: 12px;
-        color: #ffffff;
-        text-align: center;
-        border-radius: 50%;
-        -webkit-background-size: 50px auto;
-}
